a passar até ao teste 13 apenas alguns bugs de output e de verificações em testes anteriores

+++ src/Package/Services/ServicesInterface.java
@@ -7,7 +7,7 @@ public interface ServicesInterface {
     long getLatitude();
     long getLongitude();
 
-    int getPrice();
+    float getPrice();
 
     void evaluate(int star, String description,int evaluateCounter);
